+++ src/main/java/com/example/finanzasgrupo5backend/Products/Model/ProductResponse.java
@@ -13,5 +13,5 @@ public class ProductResponse {
 
     private Long id;
     private String name;
-    private Long price;
+    private Double price;
 }
